Add `dd` and `space` assembler directives
dd is similar to db, but allocations 4 bytes instead of 1 byte.

space allocates `n` number of 0's.  This allows a large area to be
statically allocated.

Changed DataArgument to pack its values in 32-bit values instead of
8-bit values.

+++ basiccpu/assembler/program.hpp
@@ -51,13 +51,13 @@ struct SymbolArgument : Argument
  */
 struct DataArgument : Argument
 {
-    std::vector<uint8_t> data;
+    std::vector<uint32_t> data;
 
-    DataArgument(const std::vector<uint8_t>& data)
+    DataArgument(const std::vector<uint32_t>& data)
     : Argument(), data(data)
     {}
 
-    DataArgument(const uint8_t* data, int length)
+    DataArgument(const uint32_t* data, int length)
     : Argument(), data(data, data+length)
     {}
 };
